fix(config): cast spectrum_max_hz to float not int in load_config
User-specified float values (e.g. 199.5) were silently truncated to int.
spectrum_max_hz is typed as float in ProcessingConfig and must stay float
through load_config so fractional Hz limits are preserved faithfully.

+++ apps/server/vibesensor/config.py
@@ -517,7 +517,7 @@ def load_config(config_path: Path | None = None) -> AppConfig:
             fft_update_hz=int(merged["processing"]["fft_update_hz"]),
             fft_n=int(merged["processing"]["fft_n"]),
             spectrum_min_hz=float(merged["processing"].get("spectrum_min_hz", 5.0)),
-            spectrum_max_hz=int(merged["processing"]["spectrum_max_hz"]),
+            spectrum_max_hz=float(merged["processing"]["spectrum_max_hz"]),
             client_ttl_seconds=int(merged["processing"].get("client_ttl_seconds", 120)),
             accel_scale_g_per_lsb=accel_scale,
         ),  # NOTE: ProcessingConfig.__post_init__ validates & clamps all fields
